Remove magic numbers for the flash page size
This replaces some values with the SPM_PAGESIZE macro from avr-libc.
Additionally, a SPM_ERASESIZE macro is defined from the makefile.  On
most devices flash can be erased one page at a time, but the
Attiny441/841 (for which this code was written) has a special 4-page
erase.  It seems avr-libc does not offer any macros to detect this, so
it is passed from the Makefile.