Generalize the linker script
Previously, a modified copy of the avr25 linker script was used.
This prevents compiling for microcontrollers that need a different
linker script, and means that the linker script might not be the latest
version, or not compatible with the current compiler version.

This commit replaces the complete linker script with an addition to the
linker script, by using the linker script INSERT feature.

Compared to the previous linker script, this removes support for the
"application" section (which was supposed to contain a default
application), since this was not being used.

To improve generality, instead of specifiying hardcoded addresses, the
bootloader size is defined in the Makefile, and passed to the linker
script.