Automatically fill the trampoline
Previously, the uploader was responsible for padding the application up
to the trampoline and putting the right jump in the trampoline,
requiring some post-processing of the generated code.

The first page received contains the reset jump, which is replaced by
the bootloader's reset jump already. Now, the bootloader takes the
original reset jump and writes it into the trampoline area.

This means that when writing the first page, an additional erase and
page write has to happen. This makes no attempt to preserve the other
contents of that page to keep the code simple.

Additionally, when a big program is uploaded that runs up to the erase
page containing the trampoline, the page is not erased a second time (to
prevent having to preserve the trampoline across the erase).

Because of these two simplifications, the bootloader now only works as
expected when writing a full program, starting at address 0 and writing
consecutive pages.