Unhardcode the reset vector instruction
Previously, this was hardcoded to 0xCBFF, or an rjmp to the start of the
bootloader at 0x1800.

However, with gcc, this first byte is not necessarily the start of the
bootloader, since there can be some data bytes before (which the reset
vector normally jumps over). In practice, the MCU would probably execute
its way through these instructions, but they could contain jumps or have
unwanted side effects.

Now, this just reads the current reset vector (which should be pointing
to the right place after a clean flash of the bootloader) and preserves
that.

Additionally, this allows moving the bootloader without this jump
breaking.