Fix the trampoline instruction
The trampoline was automatically filled with the to-be-written reset
vector. However, this is typically a RJMP instruction, which jumps
relative to the current address. If this instruction is moved, it should
also be modified to still jump to the same address.

This commit does exactly that, taking care to undo the changes for the
READ_FLASH command. This supports the RJMP and RCALL instructions,
trying to flash any other instructions at the reset vector makes the
flashing fail (which seems better than to silently fail).