Put serial_offset into PROGMEM
This saves a bit of RAM, at the cost of a bit of extra flash (more code
to load the values) in place. However, since this is the last variable
in the .data section, this avoids the overhead gcc generated for loading
data into the .data section, effectively saving 16 bytes of flash.