BootloaderTest: Remove ackLastRead
Turns out that acking the last read is actually problematic for the I²C
protocol, since then the slave does not know the transfer is completed
and might pull SDA down. This prevents the master from sending a stop
condition. In practice, this did not happen because the slave politely
withdraws from the bus when its last databyte is sent, but it is not
required to do so.