BootloaderTest: Convert to use new retval support on assertions
Previously, a patch to ArduinoUnit was needed to make it implicitly
call and return `on_failure()` on failed assertions. This patch was
rejected by the maintainer, but instead they added an explicit retval
argument to all assertions (which can be used both the return false, as
well as to print some info on failed assertions).

This commit changes all assertions that used the implicit on_failure, to
either pass on_failure explicitly, or just pass false to only return
false.