PROTOCOL: Clarify the meaning of the hardware revision field
This was intended to carry the actual hardware revision (e.g. to be
changed on every new boad version). However, doing so complicates
backward compatibility with older Devoboard firmware versions, so in
practice this revision field was left unchanged for the interface board
v1.4. This works, because the new board version is completely
compatible, so no firmware changes are needed.

This commit changes the protocol to match the meaning of the hardware
revision field to the current practice. This means that there is no
longer a way to obtain the *actual* board version, which will be fixed
in a future protocol version.