Bootloader: Refactor linker script
This slightly changes the way the linker script works. Previously, it
would prepend data into the .text output section, to put the interrupt
vectors at the start and inserting a lot of empty space to get the
trampoline and bootloader into the end of flash.

Now, puts each of these into a separate output section with a defined
starting address. This means there is no longer a ton of unused zeroes
in the output file, and using this technique it is possible to add extra
data *after* the main .text section, rather than only before (which will
be done later).

Since everything is no longer in the .text output section, the objcopy
command is modified to include all relevant sections.