Bootloader: Fix BOARD_TYPE define
Previously, this used e.g. `-DBOARD_TYPE=interfaceboard`, but it turns
out that the corresponding `#if BOARD_TYPE == interfaceboard` was always
true, even when using another value (presumably since all board types
like `interfaceboard` are undefined in the preprocessor, so resolve to
0).

This includes the board type in the macro name and checks with `defined`
instead.