bootloader: Make clean target work again
When different architectures were introduced, this command was broken,
since it was ran without any architecture or bus selected.

This commit makes clean run with all valid combinations of arch and bus,
which does some doublework and is prone to being forgotten when a new
arch or bus is added, but fixing that more properly probably requires
more extensive refactoring of the Makefil, this is probaly the simplest
solution for now.