bootloader: Make stm32/Rs485 interrupt mode work
This is unused in the bootloader, but kept here for easy code sharing
with application firmware. Support for interrupts was previously added,
but it did not actually work, because the interrupt was not enabled, and
the specific interrupt flags were not disabled again when no longer
needed (i.e. TXE would keep triggering even when there was nothing to
write).