bootloader: move some checks to compiletime
This modifies a few existing sanity checks, that essentially checked
whether the configured reply buffer size was big enough. This is
implemented using function arguments (since the max size depends on the
bus implementation so is not trivial to hardcode in a macro), which
means they are not proper compiletime constants that can be static
assert. Instead, they were checked using normal ifs, and since there was
no proper error code for them (it is not actually the master's fault)
NO_REPLY was used instead. Since that is hard to debug, and since this
is essentially a check that will always succeed or fail for a given
compilation, this commit modifies the check to produce a linker error
instead, relying on the optimizer to propagate the buffer sizes and
optimize away the linker error when the buffer is big enough.