BootloaderTest: Remove SUPPORTS_GET_MAX_PACKET_LENGTH
Originally, the plan was to let the bootloader support
GET_MAX_PACKET_LENGTH only when the max length was different from the
default, and to just not support the command and rely on the master to
fill in the default to save some bootloader flash. BootloaderTest used
the SUPPORTS_GET_MAX_PACKET_LENGTH variable to indicate this per board.
However, because the flash usage was not too much, the bootloader
instead ended up implementing the command unconditionally, making the
the variable no longer needed.

Instead of the per board variable, this now checks the protocol version,
to ensure BootloaderTest should still be able to test older bootloader
versions.