bootloader: Remove version constants from flash
These were not actually used in practice, and actually seem to have been
removed by the compiler in recent builds (older builds were not
checked).

Also, with the upcoming board info block, some of these values will no
longer be known at compiletime (except for bootloader version and board
type, but those are not the interesting ones). This means that
applications will either need to read these values from the board info
directly in the future, or the mainboard should query these values from
the bootloader and pass them to the application after it starts.