BootloaderTest: Fix downstream child select test
This updates the pin used for this test, the previously configured pin
had a pullup which seemed to prevent the test from passing (weirdly
enough this pullup seems to have been present since the first board
revisions, so it is unclear why this worked before).

In addition, this updates the testing code to enable the internal pullup
earlier, so the code should also work on pins without an external
pulldown.